requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anthropic==0.34.1
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
Authlib==1.3.2
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
claudette==0.0.7
click==8.1.7
cryptography==43.0.0
datasets==2.21.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
fastcore==1.7.1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.66.0
grpcio-health-checking==1.66.0
grpcio-tools==1.66.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
narwhals==1.5.5
numpy==2.1.0
ollama==0.3.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.23.0
protobuf==5.27.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.38.0
tenacity==8.5.0
tokenizers==0.20.0
toml==0.10.2
toolslm==0.0.5
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.33.0
weaviate-client==4.7.1
xxhash==3.5.0
yarl==1.9.4
